+++ my-first-discord-bot/commands/autoungate.js
@@ -7,11 +7,17 @@ module.exports= {
 
         /// if you dont put in any argument you will get this in return which will guide you in finding amazon ASINS
         if(!args.length){
-            return message.channel.send(`${message.author}, please input the items ASIN code from amazon. this can be found by looking at the end of the link for the sku or on the page itself! please retype this command as:
-             
-            !autoungate (Your items ASIN code minus the brackets)`);
-
-        }
+        const helpEmbed = new Discord.MessageEmbed()
+        .setColor('#00d9ff')
+        .setTitle(`You need to input an ASIN`)
+        .setAuthor('Auto Ungate Bot', 'https://www.pinclipart.com/picdir/big/54-543383_front-open-icon-free-download-png-and-cartoon.png')
+        .setDescription('\u200B')
+        .addFields(
+            {name: 'Finding Asin', value: `You can easily find the ASIN as its usually the SKU at the end of a link or in the items details.`, inline: true},
+            {name: 'Using the Command', value: 'When you have your asin redo this command but add a space and then your ASIN. easy as that!'}
+        );
+        message.channel.send(helpEmbed);
+        };
         /// concatenates the link + the asin input from the user to create a link to the product for use in the embed link
         const link = `https://sellercentral.amazon.com/product-search/search?q=${args}`