+++ my-first-discord-bot/commands/help.js
@@ -1,6 +1,6 @@
 module.exports= {
-    name: 'ping',
-    description: "this is a ping command!",
+    name: 'help',
+    description: "list of commands",
     execute(message){
         setTimeout(() => {message.delete();}, 3000)
         message.channel.send('pong').then(sentMessage => {sentMessage.delete({ timeout: 3000}); });