+++ Changelog.md
@@ -3,25 +3,37 @@
 ______________________________________
 ## [Unreleased]
 
-## [1.0.3] - 7-7-2021
+## [1.0.4] - 7-7-2021
 
 ### Added
 
 
+
 ______________________________________
+## [Released]
+
+## [1.0.3] - 7-17-2021
+
+### Added
+
+Bot will now check if the user has the required role from the PAID_ROLE variable in the .env file.
+
+You can now change your bot prefix per guild using !changeprefix as the default command and is not locked behind the paid role yet.
+
 
 ## [Released]
 
 ## [1.0.2] - 7-13-2021
 
 ### Added
-THIS IS UNSTABLE AND BARELY TESTED
+THIS IS UNSTABLE AND BARELY TESTED!!!
 
-bot will now record serverid and owner id into the Servers db and save serverid in the serverConfig db
+bot will now record serverid and owner id into the Servers db and save serverid in the serverConfig db.
 
 bot will not set default prefix for each server to ! and call upon the server id to catch the correct prefix if changed.
 
 pain
+
 
 ## [1.0.1] - 7-11-2021